+++ Common/Environs.Types.h
@@ -110,6 +110,17 @@ namespace environs {
  * */
 #define	ENVIRONS_DISPLAY_UNINITIALIZED_VALUE              		(-1)
 
+/**
+ * Environs thread states
+ */
+/** Uninitialized. Usually after creation of an Environs object. */
+#define	ENVIRONS_THREAD_NO_THREAD                         		(0)
+/** Thread is either created and not yet running or terminated. */
+#define	ENVIRONS_THREAD_DETACHEABLE                       		(1)
+/** Thread is running. */
+#define	ENVIRONS_THREAD_RUNNING                           		(2)
+
+
 
 /** Deleteable. Device object has been disabled.&nbsp; 
  * "Garbage Collection" should dispose and delete the object on the next occasion. 
@@ -843,6 +854,8 @@ namespace environs {
 #define	DEVICE_ACTIVITY_CONNECTED                         		(0x10)
 #define	DEVICE_ACTIVITY_REQUESTOR                         		(0x100)
 #define	DEVICE_ACTIVITY_RESPONDER                         		(0x200)
+#define	DEVICE_ACTIVITY_LISTENER_CLOSED                   		(0x8000)
+#define	DEVICE_ACTIVITY_PLATFORM_DISPOSED                 		(0x1000)
 	
 /**
  * Environs common native declarations