+++ Common/Environs.Types.h.cs
@@ -106,6 +106,18 @@ namespace environs
 		 * */
 		public const int ENVIRONS_DISPLAY_UNINITIALIZED_VALUE              =	(-1);
 		
+		/**
+		 * Environs thread states
+		 * Environs thread states
+		 */
+		/** Uninitialized. Usually after creation of an Environs object. */
+		public const int ENVIRONS_THREAD_NO_THREAD                         =	(0);
+		/** Thread is either created and not yet running or terminated. */
+		public const int ENVIRONS_THREAD_DETACHEABLE                       =	(1);
+		/** Thread is running. */
+		public const int ENVIRONS_THREAD_RUNNING                           =	(2);
+		
+		
 		
 		/** Deleteable. Device object has been disabled.&nbsp; 
 		 * "Garbage Collection" should dispose and delete the object on the next occasion. 
@@ -841,6 +853,8 @@ namespace environs
 		public const int DEVICE_ACTIVITY_CONNECTED                         =	(0x10);
 		public const int DEVICE_ACTIVITY_REQUESTOR                         =	(0x100);
 		public const int DEVICE_ACTIVITY_RESPONDER                         =	(0x200);
+		public const int DEVICE_ACTIVITY_LISTENER_CLOSED                   =	(0x8000);
+		public const int DEVICE_ACTIVITY_PLATFORM_DISPOSED                 =	(0x1000);
 			
 		/**
 		 * Environs common native declarations